refactored components. Button press test to-do
