+++ client/src/components/QuoteOfTheDay/QuoteOfTheDay.test.js
@@ -1,25 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import { shallow } from "enzyme";
 import ShallowRenderer from "react-test-renderer/shallow";
-import {
-    fetchKanyeQuote,
-    fetchChuckNorrisQuote,
-    fetchTaylorSwiftQuote,
-    fetchProgrammingQuote,
-    fetchAnimeQuote,
-} from "./quoteService";
 
-import QuoteOfTheDay from ".";
+import QuoteOfTheDay from "../QuoteOfTheDay";
 
 let component;
-
-function setText(text) {
-    console.log(text);
-}
-
-function setAuthor(author) {
-    console.log(author);
-}
 
 beforeEach(() => {
     component = shallow(<QuoteOfTheDay />);
@@ -31,32 +16,4 @@ test("it renders correctly", () => {
     expect(snapshotComponent).toMatchSnapshot();
 });
 
-test("It fetches Kanye West quotes correctly", () => {
-    let { text, author } = fetchKanyeQuote(setText, setAuthor);
-    expect(text !== "");
-    expect(author === "Kanye West");
-});
-
-test("It fetches Chuck Norris quotes correctly", () => {
-    let { text, author } = fetchChuckNorrisQuote(setText, setAuthor);
-    expect(text !== "");
-    expect(author === "Anonymous");
-});
-
-test("It fetches Taylor Swift quotes correctly", () => {
-    let { text, author } = fetchTaylorSwiftQuote(setText, setAuthor);
-    expect(text !== "");
-    expect(author === "Taylor Swift");
-});
-
-test("It fetches programming quotes correctly", () => {
-    let { text, author } = fetchProgrammingQuote(setText, setAuthor);
-    expect(text !== "");
-    expect(author !== "");
-});
-
-test("It fetches anime quotes correctly", () => {
-    let { text, author } = fetchAnimeQuote(setText, setAuthor);
-    expect(text !== "");
-    expect(author !== "");
-});
+// need to do tests with the button click changes
